requirements.txt
fastapi
uvicorn
langchain-community>=0.3.0
sentence-transformers
transformers>=4.41.1
torch
faiss-cpu
jinja2
unstructured[all-docs]
python-dotenv
aiofiles>=22.0,<24.0
pillow>=8.0,<11.0
